style분리, components 합체
