+++ package.json
@@ -2,6 +2,7 @@
   "name": "my-poo-diary",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -12,8 +13,10 @@
     "axios": "^1.6.3",
     "json-server": "^0.17.4",
     "json-server-auth": "^2.1.0",
+    "moment": "^2.30.1",
     "next": "14.0.4",
     "react": "^18",
+    "react-calendar": "^4.7.0",
     "react-dom": "^18",
     "react-query": "^3.39.3",
     "styled-components": "^6.1.3",
