+++ src/app/profile/model/profile.ts
@@ -1,0 +1,2 @@
+type ValuePiece = Date | null;
+export type Value = ValuePiece | [ValuePiece, ValuePiece];
